Move EVM state access into a submodule
